use laravel pint by default

+++ bin/surf.sh
@@ -6,9 +6,6 @@ ERROR='\033[91m'
 SUCCESS='\033[92m'
 RESET='\033[0m'
 
-export SURF_USER_ID=${UID}
-export SURF_GROUP_ID=$(id -g)
-
 function exit_if_containers_not_running() {
   CONTAINERS_RUNNING="$(docker-compose ps -q)"
 
@@ -30,22 +27,30 @@ if [[ "$1" == 'tls' ]]; then
     fi
 elif [[ "$1" == 'composer' ]]; then
   cd $(pwd)
-  docker-compose run --rm --no-deps laravel composer "${@:2}"
+  CONTAINERS_RUNNING="$(docker-compose ps -q)"
+
+  cd $(pwd)
+  if [[ -z "$CONTAINERS_RUNNING" ]]; then
+    docker-compose run --rm --no-deps laravel composer "${@:2}"
+  else
+    docker-compose exec -u www-data laravel composer "${@:2}"
+  fi
+
 elif [[ "$1" == 'yarn' ]]; then
   exit_if_containers_not_running
 
   cd $(pwd)
-  docker-compose exec -u ${SURF_USER_ID} laravel yarn "${@:2}"
+  docker-compose exec -u www-data laravel yarn "${@:2}"
 elif [[ "$1" == 'npx' ]]; then
   exit_if_containers_not_running
 
   cd $(pwd)
-  docker-compose exec -u ${SURF_USER_ID} laravel npx "${@:2}"
+  docker-compose exec -u www-data laravel npx "${@:2}"
 elif [[ "$1" == 'node' ]]; then
   exit_if_containers_not_running
 
   cd $(pwd)
-  docker-compose exec -u ${SURF_USER_ID} laravel node "${@:2}"
+  docker-compose exec -u www-data laravel node "${@:2}"
 elif [[ "$1" == 'aws' ]]; then
   docker-compose run --rm awscliv2 "${@:2}"
 elif [[ "$1" == 'awslocal' ]]; then
@@ -58,57 +63,57 @@ elif [[ "$1" == 'artisan' ]]; then
   exit_if_containers_not_running
 
   cd $(pwd)
-  docker-compose exec -u ${SURF_USER_ID} laravel php artisan "${@:2}"
+  docker-compose exec -u www-data laravel php artisan "${@:2}"
 elif [[ "$1" == 'publish' ]]; then
   exit_if_containers_not_running
 
   cd $(pwd)
-  docker-compose exec -u ${SURF_USER_ID} laravel php artisan larasurf:publish "${@:2}"
+  docker-compose exec -u www-data laravel php artisan larasurf:publish "${@:2}"
 elif [[ "$1" == 'config' ]]; then
   exit_if_containers_not_running
 
   cd $(pwd)
-  docker-compose exec -u ${SURF_USER_ID} laravel php artisan larasurf:config "${@:2}"
+  docker-compose exec -u www-data laravel php artisan larasurf:config "${@:2}"
 elif [[ "$1" == 'cloud-vars' ]]; then
   exit_if_containers_not_running
 
   cd $(pwd)
-  docker-compose exec -u ${SURF_USER_ID} laravel php artisan larasurf:cloud-vars "${@:2}"
+  docker-compose exec -u www-data laravel php artisan larasurf:cloud-vars "${@:2}"
 elif [[ "$1" == 'cloud-stacks' ]]; then
   exit_if_containers_not_running
 
   cd $(pwd)
-  docker-compose exec -u ${SURF_USER_ID} laravel php artisan larasurf:cloud-stacks "${@:2}"
+  docker-compose exec -u www-data laravel php artisan larasurf:cloud-stacks "${@:2}"
 elif [[ "$1" == 'cloud-emails' ]]; then
   exit_if_containers_not_running
 
   cd $(pwd)
-  docker-compose exec -u ${SURF_USER_ID} laravel php artisan larasurf:cloud-emails "${@:2}"
+  docker-compose exec -u www-data laravel php artisan larasurf:cloud-emails "${@:2}"
 elif [[ "$1" == 'cloud-ingress' ]]; then
   exit_if_containers_not_running
 
   cd $(pwd)
-  docker-compose exec -u ${SURF_USER_ID} laravel php artisan larasurf:cloud-ingress "${@:2}"
+  docker-compose exec -u www-data laravel php artisan larasurf:cloud-ingress "${@:2}"
 elif [[ "$1" == 'cloud-users' ]]; then
   exit_if_containers_not_running
 
   cd $(pwd)
-  docker-compose exec -u ${SURF_USER_ID} laravel php artisan larasurf:cloud-users "${@:2}"
+  docker-compose exec -u www-data laravel php artisan larasurf:cloud-users "${@:2}"
 elif [[ "$1" == 'cloud-domains' ]]; then
   exit_if_containers_not_running
 
   cd $(pwd)
-  docker-compose exec -u ${SURF_USER_ID} laravel php artisan larasurf:cloud-domains "${@:2}"
+  docker-compose exec -u www-data laravel php artisan larasurf:cloud-domains "${@:2}"
 elif [[ "$1" == 'cloud-images' ]]; then
   exit_if_containers_not_running
 
   cd $(pwd)
-  docker-compose exec -u ${SURF_USER_ID} laravel php artisan larasurf:cloud-images "${@:2}"
+  docker-compose exec -u www-data laravel php artisan larasurf:cloud-images "${@:2}"
 elif [[ "$1" == 'cloud-users' ]]; then
   exit_if_containers_not_running
 
   cd $(pwd)
-  docker-compose exec -u ${SURF_USER_ID} laravel php artisan larasurf:cloud-users "${@:2}"
+  docker-compose exec -u www-data laravel php artisan larasurf:cloud-users "${@:2}"
 elif [[ "$1" == 'cloud-artisan' ]] && [[ "$2" == 'tinker' ]]; then
   if [[ "$3" != '--environment' ]] || [[ "$4" != 'stage' ]] && [[ "$4" != 'production' ]]; then
     echo -e "${ERROR}Invalid environment specified${RESET}"
@@ -121,7 +126,7 @@ elif [[ "$1" == 'cloud-artisan' ]] && [[ "$2" == 'tinker' ]]; then
   echo "Waiting for task to start..."
 
   cd $(pwd)
-  TASK=$(docker-compose exec -u ${SURF_USER_ID} laravel php artisan larasurf:cloud-tasks run-for-exec --environment "$4")
+  TASK=$(docker-compose exec -u www-data laravel php artisan larasurf:cloud-tasks run-for-exec --environment "$4")
   PROJECT_NAME=$(cat larasurf.json | jq -r '."project-name"')
   PROJECT_ID=$(cat larasurf.json | jq -r '."project-id"')
   CLUSTER_NAME="larasurf-${PROJECT_ID}-$4"
@@ -137,38 +142,43 @@ elif [[ "$1" == 'cloud-artisan' ]] && [[ "$2" == 'tinker' ]]; then
     --task ${TASK}
 
   cd $(pwd)
-  docker-compose exec -u ${SURF_USER_ID} laravel php artisan larasurf:cloud-tasks stop --environment "$4" --task "${TASK}"
+  docker-compose exec -u www-data laravel php artisan larasurf:cloud-tasks stop --environment "$4" --task "${TASK}"
 elif [[ "$1" == 'cloud-artisan' ]]; then
   exit_if_containers_not_running
 
   cd $(pwd)
-  docker-compose exec -u ${SURF_USER_ID} laravel php artisan larasurf:cloud-artisan "${@:2}"
+  docker-compose exec -u www-data laravel php artisan larasurf:cloud-artisan "${@:2}"
 elif [[ "$1" == 'circleci' ]]; then
   exit_if_containers_not_running
 
   cd $(pwd)
-  docker-compose exec -u ${SURF_USER_ID} laravel php artisan larasurf:circleci "${@:2}"
+  docker-compose exec -u www-data laravel php artisan larasurf:circleci "${@:2}"
 elif [[ "$1" == 'test' ]]; then
   exit_if_containers_not_running
 
   cd $(pwd)
-  docker-compose exec -u ${SURF_USER_ID} laravel ./vendor/bin/phpunit "${@:2}"
+  docker-compose exec -u www-data laravel ./vendor/bin/phpunit "${@:2}"
+elif [[ "$1" == 'pint' ]]; then
+  exit_if_containers_not_running
+
+  cd $(pwd)
+  docker-compose exec -u www-data laravel ./vendor/bin/pint "${@:2}"
 elif [[ "$1" == 'dusk' ]]; then
   exit_if_containers_not_running
 
   cd $(pwd)
-  docker-compose exec -u ${SURF_USER_ID} -e APP_URL=http://webserver -e DUSK_DRIVER_URL=http://chrome:4444/wd/hub laravel php artisan dusk "${@:2}"
+  docker-compose exec -u www-data -e APP_URL=http://webserver -e DUSK_DRIVER_URL=http://chrome:4444/wd/hub laravel php artisan dusk "${@:2}"
 elif [[ "$1" == 'fix' ]]; then
   exit_if_containers_not_running
 
   if grep -q '"barryvdh/laravel-ide-helper"' 'composer.json'; then
     cd $(pwd)
-    docker-compose exec -u ${SURF_USER_ID} laravel bash -c 'php artisan ide-helper:generate && php artisan ide-helper:meta && php artisan ide-helper:models --write-mixin'
+    docker-compose exec -u www-data laravel bash -c 'php artisan ide-helper:generate && php artisan ide-helper:meta && php artisan ide-helper:models --write-mixin'
   fi
 
-  if grep -q '"friendsofphp/php-cs-fixer"' 'composer.json'; then
+  if grep -q '"laravel/pint"' 'composer.json'; then
     cd $(pwd)
-    docker-compose exec -u ${SURF_USER_ID} laravel ./vendor/bin/php-cs-fixer fix
+    docker-compose exec -u www-data laravel ./vendor/bin/pint
   fi
 elif [[ "$1" == 'shell' ]]; then
   exit_if_containers_not_running
@@ -178,7 +188,7 @@ elif [[ "$1" == 'shell' ]]; then
   if [[ "$2" == '--root' ]]; then
       docker-compose exec laravel bash
   else
-    docker-compose exec -u ${SURF_USER_ID} laravel bash
+    docker-compose exec -u www-data laravel bash
   fi
 elif [[ "$1" == 'fresh' ]]; then
   REFRESH_COMMAND='php artisan migrate'
@@ -216,7 +226,7 @@ elif [[ "$1" == 'fresh' ]]; then
   echo 'Database is ready!'
 
   cd $(pwd)
-  docker-compose exec -u ${SURF_USER_ID} laravel $REFRESH_COMMAND
+  docker-compose exec -u www-data laravel $REFRESH_COMMAND
 elif [[ "$1" == 'up' ]]; then
   cd $(pwd)
 
@@ -256,11 +266,11 @@ elif [[ "$1" == 'configure-new-environments' ]]; then
 
   cd $(pwd)
 
-  if ! docker-compose exec -u ${SURF_USER_ID} laravel php artisan larasurf:configure-new-environments validate-new-environments --environments "$3"; then
+  if ! docker-compose exec -u www-data laravel php artisan larasurf:configure-new-environments validate-new-environments --environments "$3"; then
     exit 1
   fi
 
-  NEW_BRANCHES=$(docker-compose exec -u ${SURF_USER_ID} laravel php artisan larasurf:configure-new-environments get-new-branches --environments "$3")
+  NEW_BRANCHES=$(docker-compose exec -u www-data laravel php artisan larasurf:configure-new-environments get-new-branches --environments "$3")
 
   cd $(pwd)
 
@@ -279,7 +289,7 @@ elif [[ "$1" == 'configure-new-environments' ]]; then
   fi
 
   cd $(pwd)
-  docker-compose exec -u ${SURF_USER_ID} laravel php artisan larasurf:configure-new-environments modify-larasurf-config --environments "$3"
+  docker-compose exec -u www-data laravel php artisan larasurf:configure-new-environments modify-larasurf-config --environments "$3"
 else
   # todo
   echo 'See: https://larasurf.com/docs'
